debug

+++ project/views.py
@@ -1,7 +1,10 @@
+import datetime
 import os
 import re
 import secrets
 import json
+from time import strptime
+
 from django.utils.timezone import now
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_http_methods
@@ -33,6 +36,12 @@ def create_project(request, user):
         return JsonResponse({'errno': 3001, 'msg': "当前用户不在该团队内"})
     if not bool(re.match("^[A-Za-z0-9][A-Za-z0-9_]{2,99}$", str(name))):
         return JsonResponse({'errno': 3002, 'msg': "项目名不合法"})
+    if estimated_start_time and estimated_end_time:
+        time_format = "%Y-%m-%d %H:%M:%S"
+        estimated_start_time = datetime.datetime.strptime(estimated_start_time, time_format)
+        estimated_end_time = datetime.datetime.strptime(estimated_end_time, time_format)
+        if estimated_start_time > estimated_end_time:
+            return JsonResponse({'errno': 3003, 'msg': "预设时间不合法"})
     team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
     new_project = Project.objects.create(project_name=name, project_description=description, project_creator=team_member, project_team=team)
     team.team_projects.add(new_project)
@@ -78,20 +87,20 @@ def show_profile(request, user):
     team_id = data_json.get('team_id')
     project_id = data_json.get('project_id')
     if not Team.objects.filter(team_id=team_id).exists():
-        return JsonResponse({'errno': 3010, 'msg': "该团队不存在"})
+        return JsonResponse({'errno': 3020, 'msg': "该团队不存在"})
     team = Team.objects.get(team_id=team_id)
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
+        return JsonResponse({'errno': 3021, 'msg': "当前用户不在该团队内"})
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3012, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3022, 'msg': "该项目不存在"})
     project = Project.objects.get(project_id=project_id)
     if project.project_recycle:
-        return JsonResponse({'errno': 3013, 'msg': '项目在回收站无法操作'})
+        return JsonResponse({'errno': 3023, 'msg': '项目在回收站无法操作'})
     project_info = project.to_json()
     project_avatar = None
     if project.project_avatar:
         project_avatar = get_avatar_base64(project.project_avatar)
-    return JsonResponse({'errno': 0, 'msg': '返回用户信息成功', 'project_info': project_info, 'project_avatar': project_avatar})
+    return JsonResponse({'errno': 0, 'msg': '返回项目信息成功', 'project_info': project_info, 'project_avatar': project_avatar})
 
 
 @csrf_exempt
@@ -109,20 +118,26 @@ def change_profile(request, user):
     estimated_end_time = data_json.get('estimated_end_time')
     recycle = data_json.get('recycle')
     if not Team.objects.filter(team_id=team_id).exists():
-        return JsonResponse({'errno': 3010, 'msg': "该团队不存在"})
+        return JsonResponse({'errno': 3030, 'msg': "该团队不存在"})
     team = Team.objects.get(team_id=team_id)
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
+        return JsonResponse({'errno': 3031, 'msg': "当前用户不在该团队内"})
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3012, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3032, 'msg': "该项目不存在"})
     team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
     project = Project.objects.get(project_id=project_id)
     if project.project_recycle:
-        return JsonResponse({'errno': 3013, 'msg': '项目在回收站无法操作'})
+        return JsonResponse({'errno': 3033, 'msg': '项目在回收站无法操作'})
     if project.project_creator != team_member:
-        return JsonResponse({'errno': 3013, 'msg': "当前用户不是该项目创建者，无法改变项目"})
+        return JsonResponse({'errno': 3034, 'msg': "当前用户不是该项目创建者，无法改变项目"})
     if not bool(re.match("^[A-Za-z0-9][A-Za-z0-9_]{2,99}$", str(name))):
-        return JsonResponse({'errno': 3002, 'msg': "项目名不合法"})
+        return JsonResponse({'errno': 3035, 'msg': "项目名不合法"})
+    if estimated_start_time and estimated_end_time:
+        time_format = "%Y-%m-%d %H:%M:%S"
+        estimated_start_time = datetime.datetime.strptime(estimated_start_time, time_format)
+        estimated_end_time = datetime.datetime.strptime(estimated_end_time, time_format)
+        if estimated_start_time > estimated_end_time:
+            return JsonResponse({'errno': 3036, 'msg': "预设时间不合法"})
     if project.project_status != 'not_started' and status == 'not_started':
         project.project_end_time = None
         project.project_start_time = None
@@ -139,7 +154,7 @@ def change_profile(request, user):
     project.project_description = description
     project.project_recycle = recycle
     project.save()
-    return JsonResponse({'errno': 0, 'msg': "项目删除成功"})
+    return JsonResponse({'errno': 0, 'msg': "项目信息修改成功"})
 
 
 @csrf_exempt
@@ -175,11 +190,11 @@ def search_status(request, user):
     team_id = data_json.get('team_id')
     status = data_json.get('status')
     if not Team.objects.filter(team_id=team_id).exists():
-        return JsonResponse({'errno': 3010, 'msg': "该团队不存在"})
+        return JsonResponse({'errno': 3040, 'msg': "该团队不存在"})
     team = Team.objects.get(team_id=team_id)
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
-    projects = Project.objects.filter(project_status=status)
+        return JsonResponse({'errno': 3041, 'msg': "当前用户不在该团队内"})
+    projects = Project.objects.filter(project_status=status, project_team=team)
     p_info = []
     for p in projects:
         if not p.project_recycle:
@@ -209,18 +224,26 @@ def create_requirement(request, user):
     estimated_start_time = data_json.get('estimated_start_time')
     estimated_end_time = data_json.get('estimated_end_time')
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3000, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3060, 'msg': "该项目不存在"})
     project = Project.objects.get(project_id=project_id)
     team = project.project_team
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3001, 'msg': "当前用户不在该团队内"})
+        return JsonResponse({'errno': 3061, 'msg': "当前用户不在该团队内"})
     if project.project_recycle:
-        return JsonResponse({'errno': 0, 'msg': "项目在回收站中，无法操作"})
+        return JsonResponse({'errno': 3062, 'msg': "项目在回收站中，无法操作"})
     if not bool(re.match("^[A-Za-z0-9][A-Za-z0-9_]{2,99}$", str(name))):
-        return JsonResponse({'errno': 3002, 'msg': "需求名不合法"})
+        return JsonResponse({'errno': 3063, 'msg': "需求名不合法"})
+    if estimated_start_time and estimated_end_time:
+        time_format = "%Y-%m-%d %H:%M:%S"
+        estimated_start_time = datetime.datetime.strptime(estimated_start_time, time_format)
+        estimated_end_time = datetime.datetime.strptime(estimated_end_time, time_format)
+        if estimated_start_time > estimated_end_time:
+            return JsonResponse({'errno': 3064, 'msg': "预设时间不合法"})
     team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
     new_requirement = Requirement.objects.create(requirement_name=name, requirement_creator=team_member, requirement_estimated_start_time=estimated_start_time, requirement_estimated_end_time=estimated_end_time)
     project.project_requirement.add(new_requirement)
+    new_requirement.requirement_project = project
+    new_requirement.save()
     return JsonResponse({'errno': 0, 'msg': "项目创建成功"})
 
 
@@ -232,20 +255,22 @@ def delete__requirement(request, user):
     project_id = data_json.get('project_id')
     requirement_id = data_json.get('requirement_id')
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3000, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3070, 'msg': "该项目不存在"})
     project = Project.objects.get(project_id=project_id)
     team = project.project_team
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
+        return JsonResponse({'errno': 3071, 'msg': "当前用户不在该团队内"})
     team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
     if project.project_recycle:
-        return JsonResponse({'errno': 0, 'msg': "项目在回收站中，无法操作"})
+        return JsonResponse({'errno': 3072, 'msg': "项目在回收站中，无法操作"})
+    if not Requirement.objects.filter(requirement_id=requirement_id).exists():
+        return JsonResponse({'errno': 3073, 'msg': "该需求不存在"})
     de_requirement = Requirement.objects.get(requirement_id=requirement_id)
-    if de_requirement.project_creator != team_member:
-        return JsonResponse({'errno': 3013, 'msg': "当前用户不是该项目创建者，无法删除项目"})
+    if de_requirement.requirement_creator != team_member:
+        return JsonResponse({'errno': 3083, 'msg': "当前用户不是该项目创建者，无法删除项目"})
     project.project_requirement.remove(de_requirement)
     de_requirement.delete()
-    return JsonResponse({'errno': 0, 'msg': "项目删除成功"})
+    return JsonResponse({'errno': 0, 'msg': "需求删除成功"})
 
 
 @csrf_exempt
@@ -256,17 +281,18 @@ def show_profile_requirement(request, user):
     project_id = data_json.get('project_id')
     requirement_id = data_json.get('requirement_id')
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3000, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3080, 'msg': "该项目不存在"})
     project = Project.objects.get(project_id=project_id)
     team = project.project_team
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
-    team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
+        return JsonResponse({'errno': 3081, 'msg': "当前用户不在该团队内"})
     if project.project_recycle:
-        return JsonResponse({'errno': 0, 'msg': "项目在回收站中，无法操作"})
+        return JsonResponse({'errno': 3082, 'msg': "项目在回收站中，无法操作"})
+    if not Requirement.objects.filter(requirement_id=requirement_id).exists():
+        return JsonResponse({'errno': 3083, 'msg': "该需求不存在"})
     requirement = Requirement.objects.get(requirement_id=requirement_id)
     requirement_info = requirement.to_json()
-    return JsonResponse({'errno': 0, 'msg': '返回用户信息成功', 'requirement_info': requirement_info})
+    return JsonResponse({'errno': 0, 'msg': '返回需求信息成功', 'requirement_info': requirement_info})
 
 
 @csrf_exempt
@@ -281,19 +307,27 @@ def change_profile_requirement(request, user):
     estimated_start_time = data_json.get('estimated_start_time')
     estimated_end_time = data_json.get('estimated_end_time')
     if not Project.objects.filter(project_id=project_id).exists():
-        return JsonResponse({'errno': 3000, 'msg': "该项目不存在"})
+        return JsonResponse({'errno': 3090, 'msg': "该项目不存在"})
     project = Project.objects.get(project_id=project_id)
     team = project.project_team
     if not TeamMember.objects.filter(tm_team_id=team, tm_user_id=user).exists():
-        return JsonResponse({'errno': 3011, 'msg': "当前用户不在该团队内"})
+        return JsonResponse({'errno': 3091, 'msg': "当前用户不在该团队内"})
     team_member = TeamMember.objects.get(tm_team_id=team, tm_user_id=user)
     if project.project_recycle:
-        return JsonResponse({'errno': 0, 'msg': "项目在回收站中，无法操作"})
+        return JsonResponse({'errno': 3092, 'msg': "项目在回收站中，无法操作"})
+    if not Requirement.objects.filter(requirement_id=requirement_id).exists():
+        return JsonResponse({'errno': 3093, 'msg': "该需求不存在"})
     requirement = Requirement.objects.get(requirement_id=requirement_id)
-    if requirement.project_creator != team_member:
-        return JsonResponse({'errno': 3013, 'msg': "当前用户不是该项目创建者，无法删除项目"})
+    if requirement.requirement_creator != team_member:
+        return JsonResponse({'errno': 3094, 'msg': "当前用户不是该项目创建者，无法删除项目"})
     if not bool(re.match("^[A-Za-z0-9][A-Za-z0-9_]{2,99}$", str(name))):
-        return JsonResponse({'errno': 3002, 'msg': "需求名不合法"})
+        return JsonResponse({'errno': 3095, 'msg': "需求名不合法"})
+    if estimated_start_time and estimated_end_time:
+        time_format = "%Y-%m-%d %H:%M:%S"
+        estimated_start_time = datetime.datetime.strptime(estimated_start_time, time_format)
+        estimated_end_time = datetime.datetime.strptime(estimated_end_time, time_format)
+        if estimated_start_time > estimated_end_time:
+            return JsonResponse({'errno': 3096, 'msg': "预设时间不合法"})
     if requirement.requirement_status != 'not_started' and status == 'not_started':
         requirement.requirement_end_time = None
         requirement.requirement_start_time = None
@@ -307,4 +341,4 @@ def change_profile_requirement(request, user):
     requirement.requirement_estimated_start_time = estimated_start_time
     requirement.requirement_name = name
     requirement.save()
-    return JsonResponse({'errno': 0, 'msg': "项目删除成功"})
+    return JsonResponse({'errno': 0, 'msg': "需求信息修改成功"})
